Add the nuget-base plugin
