+++ src/main/resources/META-INF/gradle-plugins/nunit-base.properties
@@ -1,0 +1,1 @@
+implementation-class=com.ullink.gradle.nunit.NUnitBasePlugin
